Changes:
- Cleanup Model inheritance
