+++ edgy/core/db/fields/types.py
@@ -119,7 +119,7 @@ class BaseFieldType(BaseFieldDefinitions, ABC):
         value: Any,
     ) -> dict[str, Any]:
         """
-        Inverse of clean. Transforms column(s) to a field for a pydantic model (EdgyBaseModel).
+        Inverse of clean. Transforms column(s) to a field for edgy.Model.
         Validation happens later.
 
         Args:
