+++ edgy/core/db/models/mixins/__init__.py
@@ -1,5 +1,6 @@
+from .db import DatabaseMixin
 from .generics import DeclarativeMixin
 from .reflection import ReflectedModelMixin
 from .row import ModelRowMixin
 
-__all__ = ["DeclarativeMixin", "ModelRowMixin", "ReflectedModelMixin"]
+__all__ = ["DeclarativeMixin", "ModelRowMixin", "ReflectedModelMixin", "DatabaseMixin"]
